requirements.txt
aiogram 
aiohttp
openai
aioboto3
Pillow
python-dotenv
numpy
opencv-python
pytesseract
requests
openai
pyzbar
imutils